Draw Line

+++ SoftRenderer/Vector4.cpp
@@ -93,7 +93,7 @@ Vector4& Vector4::operator*=(float k)
 
 Vector4& Vector4::operator/=(float k)
 {
-	float oneOver = 1.0 / k;
+	float oneOver = 1.0f / k;
 	x *= oneOver;
 	y *= oneOver;
 	z *= oneOver;
@@ -106,7 +106,7 @@ void Vector4::Normalize()
 	float sq = x * x + y * y + z * z;
 	if (sq > 0.0)
 	{
-		double oneOver = 1.0 / sqrt(sq);
+		float oneOver = 1.0 / sqrt(sq);
 		x *= oneOver;
 		y *= oneOver;
 		z *= oneOver;
